requirements.txt
langchain==0.2.6
langchain-core==0.2.10
langchain-groq==0.1.5
pydantic==1.10.13
python-dotenv==1.0.1
